Move the Settings trait to a separate package
