+++ rust/Cargo.toml
@@ -4,5 +4,6 @@ members = [
   "agama-dbus-server",
   "agama-derive",
   "agama-lib",
-  "agama-locale-data"
+  "agama-locale-data",
+  "agama-settings"
 ]
