+++ rust/agama-cli/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2021"
 [dependencies]
 clap = { version = "4.1.4", features = ["derive"] }
 agama-lib = { path="../agama-lib" }
+agama-settings = { path="../agama-settings" }
 serde = { version = "1.0.152" }
 serde_json = "1.0.91"
 serde_yaml = "0.9.17"
