+++ rust/agama-cli/src/config.rs
@@ -2,8 +2,8 @@ use crate::error::CliError;
 use crate::printers::{print, Format};
 use agama_lib::connection;
 use agama_lib::install_settings::{InstallSettings, Scope};
-use agama_lib::settings::{SettingObject, SettingValue, Settings};
 use agama_lib::Store as SettingsStore;
+use agama_settings::{settings::Settings, SettingObject, SettingValue};
 use clap::Subcommand;
 use convert_case::{Case, Casing};
 use std::str::FromStr;
