+++ rust/agama-lib/src/install_settings.rs
@@ -1,7 +1,7 @@
 //! Configuration settings handling
 //!
 //! This module implements the mechanisms to load and store the installation settings.
-use crate::settings::Settings;
+use agama_settings::settings::Settings;
 use crate::{
     network::NetworkSettings, software::SoftwareSettings, storage::StorageSettings,
     users::UserSettings,
