+++ rust/agama-lib/src/users/settings.rs
@@ -1,4 +1,4 @@
-use crate::settings::Settings;
+use agama_settings::settings::Settings;
 use agama_derive::Settings;
 use serde::{Deserialize, Serialize};
