+++ rust/agama-settings/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "agama-settings"
+version = "1.0.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+agama-derive = { path="../agama-derive" }
+thiserror = "1.0.43"
