+++ rust/agama-settings/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod error;
+pub mod settings;
+
+pub use self::error::SettingsError;
+pub use self::settings::{SettingObject, SettingValue};
